requirements.txt
asgiref==3.7.0
beautifulsoup4==4.12.2
boto3==1.24.22
botocore==1.27.22
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2023.1
distlib==0.3.8
Django==5.0
django-bootstrap4==23.2
django-crispy-forms==2.1
django-environ==0.11.2
django-filter==23.5
django-localflavor==4.0
django-storages==1.14.2
environ==1.0
filelock==3.13.1
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-stdnum==1.19
requests==2.31.0
s3transfer==0.6.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.9
virtualenv==20.25.0
waitress==2.1.2
whitenoise==6.6.0

